# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XXXXXXXXXXX",
        "title": "Future对外开放API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/ws/send": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发送消息"
                ],
                "parameters": [
                    {
                        "description": "input",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.SendMessage"
                        }
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.SendMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "to_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}